Disallow ending first word in PR title in "s"
This just further restricts our PR titles, which ideally would all be
the imperative, present tense (like a good Git commit[1]). While we
can't really do that in regex, we can do a little better.

One minor side-effect of this change is that because we have a literal
" " character in the regex, we're now requiring multiple words in a PR -
I doubt there's a downside to this, though.

The regex is now basically:

* One capitalized letter
* Any number of word characters
* End first word not with "s"
* Space character
* Any characters
* Not end in punctuation

From tpope:

> Write your commit message in the imperative: "Fix bug" and not "Fixed bug"
  or "Fixes bug."  This convention matches up with commit messages generated
  by commands like git merge and git revert.

[1]: https://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html